+++ Resteurent/WSRestaurant/Reposetory/OrderReposetory.cs
@@ -1,6 +1,7 @@
-﻿using System.Data;
+﻿using System.Collections.Generic;
+using System.Data;
+using System.Diagnostics.Metrics;
 using LiolamResteurent;
-
 namespace WSRestaurant
 {
     public class OrderRerposetory : Reposetory, IReposetory<Orders>
@@ -8,16 +9,52 @@ namespace WSRestaurant
         public OrderRerposetory(DBContext dbContext) : base(dbContext) { }
         public bool create(Orders model)
         {
-            string sql = $@"INSERT INTO Orders (OrderDate) VALUES (@OrderDate)";
-            this.dbContext.AddParameter("@OrderDescription", model.OrderDate.ToString());
-            return this.dbContext.Insert(sql);
+            List<int> counts = new List<int>();
+            List<Dishes> dishes = new List<Dishes>();
+            string sql = $@"INSERT INTO Orders (CustomerId, OrderDate) VALUES (@CustomerId, @OrderDate)";
+            this.dbContext.AddParameter("@CustomerId", model.Customer.Id);
+            this.dbContext.AddParameter("@OrderDate", model.OrderDate.ToString());
+            bool ok = this.dbContext.Insert(sql);
+            foreach(Dishes dish in model.dishes)
+            {
+                if (dishes.IndexOf(dish) == -1)
+                {
+                    dishes.Add(dish);
+                    counts.Add(model.dishes.Count(d => d.Equals(dish)));
+                }
+            }
+            if (ok)
+            {
+                int i = 0;
+                foreach (Dishes dish in dishes)//TO fix
+                {
+                    sql = $@"INSERT INTO DishOrder (DishId, OrderId, Price, Quantity) VALUES (@DishId, @OrderId, @Price, @Quantity)";
+                    this.dbContext.AddParameter("@DishId", dish.Id);
+                    this.dbContext.AddParameter("@OrderDate", model.OrderDate.ToString());
+                    this.dbContext.AddParameter("@Price", dish.DishPrice.ToString());
+                    this.dbContext.AddParameter("@Quantity", counts.ToArray()[i].ToString());
+                    if(!this.dbContext.Insert(sql))
+                    {
+                        throw new Exception("return false seconed");
+                    }
+                    i++;
+                }
+                return ok;
+            }
+            else throw new Exception("return false");
         }
 
         public bool delete(string id)
         {
-            string sql = $@"DELETE FROM Orders WHERE OrderId=@OrderId";
+            string sql = $@"DELETE DishOrder WHERE OrderId=@OrderId";
             this.dbContext.AddParameter("@OrderId", id);
-            return this.dbContext.Delete(sql);
+            if (this.dbContext.Delete(sql))
+            {
+                sql = $@"DELETE FROM Orders WHERE OrderId=@OrderId";
+                this.dbContext.AddParameter("@OrderId", id);
+                return this.delete(sql);
+            }
+            else throw new Exception("return false");
         }
 
         public List<Orders> getAll()
@@ -45,12 +82,63 @@ namespace WSRestaurant
                 return this.modelFactory.createOrderObject.CreateModel(dataReader);
             }
         }
-        public bool update(Orders model)
+
+        public Orders getByCustomer(string customerId)
         {
-            string sql = $@"UPDATE Orders SET OrderDate = @OrderDate WHERE OrderId == @OrderId";
-            this.dbContext.AddParameter("@OrderName", model.OrderDate.ToString());
-            this.dbContext.AddParameter("@OrderId", model.Id);
-            return this.dbContext.Update(sql);
+            string sql = "SELECT FROM Orders WHERE CustomerId = @CustomerId";
+            this.dbContext.AddParameter("@CustomerId", customerId);
+            using (IDataReader dataReader = this.dbContext.Read(sql))
+            {
+                dataReader.Read();
+                return this.modelFactory.createOrderObject.CreateModel(dataReader);
+            }
+        }
+        public bool update(Orders model) //not in use
+        {
+            List<int> counts = new List<int>();
+            List<Dishes> dishes = new List<Dishes>();
+            string sql = $@"UPDATE Orders SET CustomerId = @CustomerId, OrderDate = @OrderDate WHERE OrderId == @OrderId";
+            this.dbContext.AddParameter("@CustomerId", model.Customer.Id);
+            this.dbContext.AddParameter("@OrderDate", model.OrderDate.ToString());
+            bool ok = this.dbContext.Update(sql);
+            foreach (Dishes dish in model.dishes)
+            {
+                if (dishes.IndexOf(dish) == -1)
+                {
+                    dishes.Add(dish);
+                    counts.Add(model.dishes.Count(d => d.Equals(dish)));
+                }
+            }
+            if (ok)
+            {
+                int i = 0;
+                foreach (Dishes dish in dishes)//TO fix
+                {
+                    sql = $@"UPDATE DishOrder OrderId = @OrderId, Price = @Price, Quantity = @Quantity WHERE (SELECT OrderId FROM DishOrder WHERE ORDER BY DishId LIMIT 1) = @DishId";
+                    this.dbContext.AddParameter("@DishId", dish.Id);
+                    this.dbContext.AddParameter("@OrderDate", model.OrderDate.ToString());
+                    this.dbContext.AddParameter("@Price", dish.DishPrice.ToString());
+                    this.dbContext.AddParameter("@Quantity", counts.ToArray()[i].ToString());
+                    i++;
+                }
+                return ok;
+            }
+            else throw new Exception("return false");
+
+        }
+
+        public bool deleteByCustomer(string customerId)
+        {
+            Orders order = this.getByCustomer(customerId);
+            string sql = $@"DELETE FROM Orders WHERE CustomerId=@CustomerId";
+            this.dbContext.AddParameter("@CustomerId", customerId);
+            if (this.dbContext.Delete(sql))
+            {
+                sql = $@"DELETE FROM DishOrder WHERE OrderId=@OrderId";
+                this.dbContext.AddParameter("@DishId", order.Id);
+                return this.dbContext.Delete(sql);
+            }
+            else throw new Exception("return false");
         }
     }
 }
